+++ python/paddle/fluid/tests/unittests/ir/inference/test_trt_deformable_conv.py
@@ -16,6 +16,7 @@ import os
 import unittest
 import numpy as np
 from inference_pass_test import InferencePassTest
+import paddle
 import paddle.fluid as fluid
 import paddle.fluid.core as core
 from paddle.fluid.core import PassVersionChecker
@@ -38,7 +39,7 @@ class TRTDeformableConvTest(InferencePassTest):
                 name='mask', shape=self.mask_size, dtype=self.dtype
             )
 
-            output = fluid.layers.deformable_conv(
+            output = paddle.static.nn.common.deformable_conv(
                 input,
                 offset,
                 mask,
