+++ python/paddle/fluid/tests/unittests/test_deformable_conv_op.py
@@ -15,7 +15,6 @@
 import paddle
 import unittest
 import numpy as np
-import paddle.fluid as fluid
 from op_test import OpTest
 from paddle.fluid.framework import _test_eager_guard
 
@@ -395,13 +394,13 @@ class TestModulatedDeformableConvInvalidInput(unittest.TestCase):
         def test_invalid_input():
             paddle.enable_static()
             input = [1, 3, 32, 32]
-            offset = fluid.data(
+            offset = paddle.static.data(
                 name='offset', shape=[None, 3, 32, 32], dtype='float32'
             )
-            mask = fluid.data(
+            mask = paddle.static.data(
                 name='mask', shape=[None, 3, 32, 32], dtype='float32'
             )
-            loss = fluid.layers.deformable_conv(
+            loss = paddle.static.nn.common.deformable_conv(
                 input, offset, mask, num_filters=4, filter_size=1
             )
 
@@ -409,16 +408,16 @@ class TestModulatedDeformableConvInvalidInput(unittest.TestCase):
 
         def test_invalid_offset():
             paddle.enable_static()
-            input = fluid.data(
+            input = paddle.static.data(
                 name='input', shape=[None, 3, 32, 32], dtype='int32'
             )
-            offset = fluid.data(
+            offset = paddle.static.data(
                 name='offset', shape=[None, 3, 32, 32], dtype='float32'
             )
-            mask = fluid.data(
+            mask = paddle.static.data(
                 name='mask', shape=[None, 3, 32, 32], dtype='float32'
             )
-            loss = fluid.layers.deformable_conv(
+            loss = paddle.static.nn.common.deformable_conv(
                 input, offset, mask, num_filters=4, filter_size=1
             )
 
@@ -426,16 +425,16 @@ class TestModulatedDeformableConvInvalidInput(unittest.TestCase):
 
         def test_invalid_filter():
             paddle.enable_static()
-            input = fluid.data(
+            input = paddle.static.data(
                 name='input_filter', shape=[None, 3, 32, 32], dtype='float32'
             )
-            offset = fluid.data(
+            offset = paddle.static.data(
                 name='offset_filter', shape=[None, 3, 32, 32], dtype='float32'
             )
-            mask = fluid.data(
+            mask = paddle.static.data(
                 name='mask_filter', shape=[None, 3, 32, 32], dtype='float32'
             )
-            loss = fluid.layers.deformable_conv(
+            loss = paddle.static.nn.common.deformable_conv(
                 input, offset, mask, num_filters=4, filter_size=0
             )
