+++ python/paddle/fluid/tests/unittests/test_deformable_conv_v1_op.py
@@ -15,9 +15,7 @@
 import paddle
 import unittest
 import numpy as np
-import paddle.fluid as fluid
 from op_test import OpTest
-from paddle.fluid.framework import _test_eager_guard
 
 
 def dmc_bilinear(data_im, height, width, h, w):
@@ -346,47 +344,6 @@ class TestWithGroup(TestModulatedDeformableConvOp):
 class TestWithDouble(TestModulatedDeformableConvOp):
     def init_type(self):
         self.dtype = np.float64
-
-
-class TestModulatedDeformableConvV1InvalidInput(unittest.TestCase):
-    def test_error(self):
-        def test_invalid_input():
-            input = [1, 3, 32, 32]
-            offset = fluid.data(
-                name='offset', shape=[None, 3, 32, 32], dtype='float32'
-            )
-            loss = fluid.layers.deformable_conv(
-                input,
-                offset,
-                mask=None,
-                num_filters=4,
-                filter_size=1,
-                modulated=False,
-            )
-
-        self.assertRaises(TypeError, test_invalid_input)
-
-        def test_invalid_offset():
-            input = fluid.data(
-                name='input', shape=[None, 3, 32, 32], dtype='int32'
-            )
-            offset = fluid.data(
-                name='offset', shape=[None, 3, 32, 32], dtype='float32'
-            )
-            loss = fluid.layers.deformable_conv(
-                input,
-                offset,
-                mask=None,
-                num_filters=4,
-                filter_size=1,
-                modulated=False,
-            )
-
-        self.assertRaises(TypeError, test_invalid_offset)
-
-    def test_error_with_eager_guard(self):
-        with _test_eager_guard():
-            self.test_error()
 
 
 if __name__ == '__main__':
