+++ python/paddle/fluid/tests/unittests/test_layers.py
@@ -4797,57 +4797,6 @@ class TestBook(LayerTest):
             self.assertIsNotNone(out)
         print(str(program))
 
-    def test_deformable_conv(self):
-        with self.static_graph():
-            input = layers.data(
-                name='input',
-                append_batch_size=False,
-                shape=[2, 3, 32, 32],
-                dtype="float32",
-            )
-            offset = layers.data(
-                name='offset',
-                append_batch_size=False,
-                shape=[2, 18, 32, 32],
-                dtype="float32",
-            )
-            mask = layers.data(
-                name='mask',
-                append_batch_size=False,
-                shape=[2, 9, 32, 32],
-                dtype="float32",
-            )
-            out = layers.deformable_conv(
-                input=input,
-                offset=offset,
-                mask=mask,
-                num_filters=2,
-                filter_size=3,
-                padding=1,
-            )
-            return out
-
-    def test_deformable_conv2(self):
-        with self.static_graph():
-            input = fluid.data(
-                name='input', shape=[None, 3, None, None], dtype="float32"
-            )
-            offset = fluid.data(
-                name='offset', shape=[None, 18, None, None], dtype="float32"
-            )
-            mask = fluid.data(
-                name='mask', shape=[None, 9, None, None], dtype="float32"
-            )
-            out = layers.deformable_conv(
-                input=input,
-                offset=offset,
-                mask=mask,
-                num_filters=2,
-                filter_size=3,
-                padding=1,
-            )
-            return out
-
     def test_unfold(self):
         with self.static_graph():
             x = layers.data(name='x', shape=[3, 20, 20], dtype='float32')
@@ -4899,33 +4848,6 @@ class TestBook(LayerTest):
                 trans_std=0.1,
             )
         return out
-
-    def test_deformable_conv_v1(self):
-        with program_guard(
-            fluid.default_main_program(), fluid.default_startup_program()
-        ):
-            input = layers.data(
-                name='input',
-                append_batch_size=False,
-                shape=[2, 3, 32, 32],
-                dtype="float32",
-            )
-            offset = layers.data(
-                name='offset',
-                append_batch_size=False,
-                shape=[2, 18, 32, 32],
-                dtype="float32",
-            )
-            out = layers.deformable_conv(
-                input=input,
-                offset=offset,
-                mask=None,
-                num_filters=2,
-                filter_size=3,
-                padding=1,
-                modulated=False,
-            )
-            return out
 
     def test_retinanet_target_assign(self):
         with program_guard(
